Chore: delete unused components
