+++ services/frontend/src/constants/date.ts
@@ -4,7 +4,7 @@ export const DISPLAY_DATETIME_FORMAT = 'DD.MM.YYYY HH:mm'
 export const DISPLAY_DATE_FORMAT_DEV = 'DD.MM.YYYY HH:mm:ss'
 export const LONG_DATE_TIME_FORMAT = 'D MMMM YYYY [at] HH:mm:ss'
 export const ISO_DATE_FORMAT = 'YYYY-MM-DD'
-export const ISO_DATE_FORMAT_DEV = 'YYYY-MM-DD HH:mm:ss'
+// export const ISO_DATE_FORMAT_DEV = 'YYYY-MM-DD HH:mm:ss' // uncomment if needed
 export const YEAR_DATE_FORMAT = 'YYYY'
 
 export enum DateFormat {
